% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{amphi_load_packages}
\alias{amphi_load_packages}
\title{Load and install essential dependencies for 'salutUNEDstats'}
\usage{
amphi_load_packages(update_packages = FALSE)
}
\arguments{
\item{update_packages}{Logical. If \code{TRUE}, it attempts to update all packages
to their latest version. Defaults to \code{FALSE}.}
}
\value{
No return value, but loads the packages into the global environment.
}
\description{
This function uses the 'pacman' package to check, install (if necessary),
and load the key dependencies of the 'salutUNEDstats' package for solving the exercises.
}
